Code review: Apply automated fixes
Automated fixes applied via fix-code-issues.sh script:

1. Code Formatting
   - Formatted all Go files with gofmt
   - 67 files reformatted for consistency

2. Critical Fixes
   - Fixed lock value copy in daemon/queue/queue.go
     * GetMetrics() now returns pointer to copy without mutex
     * Eliminates go vet warning about copying sync.RWMutex

   - Fixed build tags in providers/hyperv/client_test.go
     * Moved to proper //go:build syntax at file top

3. Cleanup
   - Removed unused variables:
     * providers/aws/client.go: exportResult
     * providers/azure/export.go: containerURLParsed
     * examples/migration_orchestrator_example.go: log import

4. Type Fixes
   - Fixed int to int64 conversions for progress reporters:
     * providers/aws/export.go
     * providers/gcp/export.go

5. Dependencies
   - Updated go.mod and tidied dependencies

Remaining Issues (Manual Fix Required):
- AWS SDK v2 API mismatches (providers/aws/)
- Azure SDK API mismatches (providers/azure/)
- GCP SDK undefined types (providers/gcp/)
- Test mock interface mismatches
- Integration test compilation errors

See CODE_REVIEW_REPORT.md for complete analysis and next steps.

+++ providers/gcp/export.go
@@ -34,7 +34,7 @@ func (c *Client) ExportImageToGCS(ctx context.Context, imageName, bucket, output
 		Project: c.config.ProjectID,
 		Image:   imageName,
 		ImageExportRequest: &computepb.ImageExportRequest{
-			DestinationUri: &gcsURI,
+			DestinationUri:  &gcsURI,
 			DiskImageFormat: makeStringPtr("vmdk"),
 		},
 	}
@@ -73,13 +73,13 @@ func (c *Client) ExportImageToGCS(ctx context.Context, imageName, bucket, output
 	c.logger.Info("VMDK downloaded successfully", "path", localPath, "size_bytes", size)
 
 	return &ExportResult{
-		ImageName:  imageName,
-		Format:     "vmdk",
-		LocalPath:  localPath,
-		Size:       size,
-		GCSBucket:  bucket,
-		GCSObject:  fmt.Sprintf("%s.vmdk", imageName),
-		GCSURI:     gcsURI,
+		ImageName: imageName,
+		Format:    "vmdk",
+		LocalPath: localPath,
+		Size:      size,
+		GCSBucket: bucket,
+		GCSObject: fmt.Sprintf("%s.vmdk", imageName),
+		GCSURI:    gcsURI,
 	}, nil
 }
 
@@ -366,7 +366,7 @@ func (pr *progressReader) Read(p []byte) (int, error) {
 
 	if pr.reporter != nil && pr.total > 0 {
 		percentage := int((pr.current * 100) / pr.total)
-		pr.reporter.Update(percentage)
+		pr.reporter.Update(int64(percentage))
 	}
 
 	return n, err
@@ -374,15 +374,15 @@ func (pr *progressReader) Read(p []byte) (int, error) {
 
 // ExportResult contains the result of a GCS export operation
 type ExportResult struct {
-	ImageName  string
-	DiskName   string
-	DiskType   string // "boot", "data-1", "data-2", etc.
-	Format     string
-	LocalPath  string
-	Size       int64
-	GCSBucket  string
-	GCSObject  string
-	GCSURI     string
+	ImageName string
+	DiskName  string
+	DiskType  string // "boot", "data-1", "data-2", etc.
+	Format    string
+	LocalPath string
+	Size      int64
+	GCSBucket string
+	GCSObject string
+	GCSURI    string
 }
 
 // CreateExportManifest creates a JSON manifest file for multi-disk exports
